requirements.txt
paho-mqtt~=1.5.0
pymssql~=2.1.4